Refactor:  contexts로 폴더명 수정
issue #13
