+++ src/components/orderArea/MoneySlot.jsx
@@ -3,7 +3,7 @@ import React, { useContext, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { Container, MoneyInput, Unit } from 'components/orderArea/MoneySlot.style';
 import { addCommasToNumber } from 'utils/util';
-import { FinalPayContext } from 'Context/FinalPayProvider';
+import { FinalPayContext } from 'contexts/FinalPayProvider';
 
 export default function MoneySlot({ inputPay, updateInputPay, resetInputPay }) {
   const finalPay = useContext(FinalPayContext);
