+++ src/components/orderArea/ProductHole.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { Container, Product } from 'components/orderArea/ProductHole.style';
-import { SelectedProductContext } from 'Context/SelectedProductProvider';
+import { SelectedProductContext } from 'contexts/SelectedProductProvider';
 
 export default function ProductHole() {
   const selectedProduct = useContext(SelectedProductContext);
