+++ src/pages/VendingMachine.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ProductsArea from 'components/productsArea/ProductsArea';
 import OrderArea from 'components/orderArea/OrderArea';
 import { Container } from 'pages/VendingMachine.style';
-import { FinalPayProvider } from 'Context/FinalPayProvider';
+import { FinalPayProvider } from 'contexts/FinalPayProvider';
 
 export default function VendingMachine() {
   return (
